refactor: 이미지용 컨버터 수정 - 비트맵을 변환하는 코드를 컨버터에 추가
