+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/mypage/EditPetActivity.kt
@@ -17,7 +17,7 @@ import com.bumptech.glide.Glide
 import com.nbcfinalteam2.ddaraogae.R
 import com.nbcfinalteam2.ddaraogae.databinding.ActivityEditPetBinding
 import com.nbcfinalteam2.ddaraogae.presentation.ui.model.DogItemModel
-import com.nbcfinalteam2.ddaraogae.presentation.util.UriToByteArrayConverter.uriToByteArray
+import com.nbcfinalteam2.ddaraogae.presentation.util.ImageConverter.uriToByteArray
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
